+++ spec/client/realm_client_spec.rb
@@ -64,7 +64,7 @@ RSpec.describe KeycloakAdmin::RealmClient do
       allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:faraday_options).and_return faraday_options
 
       expect(KeycloakAdmin::Resource).to receive(:new).with(
-        "http://auth.service.io/auth/admin/realms", faraday_options).and_call_original
+        "http://auth.service.io/auth/admin/realms", faraday_options, anything).and_call_original
 
       realms = @realm_client.list
       expect(realms.length).to eq 1
@@ -91,7 +91,7 @@ RSpec.describe KeycloakAdmin::RealmClient do
       allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:faraday_options).and_return faraday_options
 
       expect(KeycloakAdmin::Resource).to receive(:new).with(
-        "http://auth.service.io/auth/admin/realms/valid-realm", faraday_options).and_call_original
+        "http://auth.service.io/auth/admin/realms/valid-realm", faraday_options, anything).and_call_original
 
       expect(@realm_client.delete).to be_truthy
     end
@@ -121,7 +121,7 @@ RSpec.describe KeycloakAdmin::RealmClient do
       allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:faraday_options).and_return faraday_options
 
       expect(KeycloakAdmin::Resource).to receive(:new).with(
-        "http://auth.service.io/auth/admin/realms", faraday_options).and_call_original
+        "http://auth.service.io/auth/admin/realms", faraday_options, anything).and_call_original
 
       @realm_client.save(realm)
     end
@@ -147,7 +147,7 @@ RSpec.describe KeycloakAdmin::RealmClient do
       allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:faraday_options).and_return faraday_options
 
       expect(KeycloakAdmin::Resource).to receive(:new).with(
-        "http://auth.service.io/auth/admin/realms/valid-realm", faraday_options).and_call_original
+        "http://auth.service.io/auth/admin/realms/valid-realm", faraday_options, anything).and_call_original
 
       @realm_client.update(realm_json)
     end